@@ -2,7 +2,6 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-from __future__ import print_function
 
 import ntpath
 import os
@@ -25,8 +24,6 @@ import gyp.MSVSUtil as MSVSUtil
 import gyp.MSVSVersion as MSVSVersion
 from gyp.common import GypError
 from gyp.common import OrderedSet
-
-PY3 = bytes != str
 
 
 # Regular expression for validating Visual Studio GUIDs.  If the GUID
@@ -84,7 +81,6 @@ generator_additional_non_configuration_keys = [
     "msvs_enable_winrt",
     "msvs_requires_importlibrary",
     "msvs_enable_winphone",
-    "msvs_enable_marmasm",
     "msvs_application_type_revision",
     "msvs_target_platform_version",
     "msvs_target_platform_minversion",
@@ -121,9 +117,7 @@ def _GetDomainAndUserName():
             call = subprocess.Popen(
                 ["net", "config", "Workstation"], stdout=subprocess.PIPE
             )
-            config = call.communicate()[0]
-            if PY3:
-                config = config.decode("utf-8")
+            config = call.communicate()[0].decode("utf-8")
             username_re = re.compile(r"^User name\s+(\S+)", re.MULTILINE)
             username_match = username_re.search(config)
             if username_match:
@@ -183,7 +177,8 @@ def _FixPath(path):
 
 def _IsWindowsAbsPath(path):
     """
-  On Cygwin systems Python needs a little help determining if a path is an absolute Windows path or not, so that
+  On Cygwin systems Python needs a little help determining if a path
+  is an absolute Windows path or not, so that
   it does not treat those as relative, which results in bad paths like:
   '..\\C:\\<some path>\\some_source_code_file.cc'
   """
@@ -319,7 +314,7 @@ def _ConfigBaseName(config_name, platform_name):
 
 def _ConfigFullName(config_name, config_data):
     platform_name = _ConfigPlatform(config_data)
-    return "%s|%s" % (_ConfigBaseName(config_name, platform_name), platform_name)
+    return "{}|{}".format(_ConfigBaseName(config_name, platform_name), platform_name)
 
 
 def _ConfigWindowsTargetPlatformVersion(config_data, version):
@@ -423,13 +418,7 @@ def _BuildCommandLineForRuleRaw(
         # file out of the raw command string, and some commands (like python) are
         # actually batch files themselves.
         command.insert(0, "call")
-        # Fix the paths
-        # TODO(quote): This is a really ugly heuristic, and will miss path fixing
-        #              for arguments like "--arg=path" or "/opt:path".
-        # If the argument starts with a slash or dash, it's probably a command line
-        # switch
-        arguments = [i if (i[:1] in "/-") else _FixPath(i) for i in cmd[1:]]
-        arguments = [i.replace("$(InputDir)", "%INPUTDIR%") for i in arguments]
+        arguments = [i.replace("$(InputDir)", "%INPUTDIR%") for i in cmd[1:]]
         arguments = [MSVSSettings.FixVCMacroSlashes(i) for i in arguments]
         if quote_cmd:
             # Support a mode for using cmd directly.
@@ -620,7 +609,7 @@ def _GenerateNativeRulesForMSVS(p, rules, output_dir, spec, options):
     spec: the project dict
     options: global generator options
   """
-    rules_filename = "%s%s.rules" % (spec["target_name"], options.suffix)
+    rules_filename = "{}{}.rules".format(spec["target_name"], options.suffix)
     rules_file = MSVSToolFile.Writer(
         os.path.join(output_dir, rules_filename), spec["target_name"]
     )
@@ -666,7 +655,7 @@ def _GenerateExternalRules(rules, output_dir, spec, sources, options, actions_to
     options: global generator options
     actions_to_add: The list of actions we will add to.
   """
-    filename = "%s_rules%s.mk" % (spec["target_name"], options.suffix)
+    filename = "{}_rules{}.mk".format(spec["target_name"], options.suffix)
     mk_file = gyp.common.WriteOnDiff(os.path.join(output_dir, filename))
     # Find cygwin style versions of some paths.
     mk_file.write('OutDirCygwin:=$(shell cygpath -u "$(OutDir)")\n')
@@ -709,7 +698,7 @@ def _GenerateExternalRules(rules, output_dir, spec, sources, options, actions_to
             cmd = ['"%s"' % i for i in cmd]
             cmd = " ".join(cmd)
             # Add it to the makefile.
-            mk_file.write("%s: %s\n" % (" ".join(outputs), " ".join(inputs)))
+            mk_file.write("{}: {}\n".format(" ".join(outputs), " ".join(inputs)))
             mk_file.write("\t%s\n\n" % cmd)
     # Close up the file.
     mk_file.close()
@@ -753,7 +742,7 @@ def _EscapeEnvironmentVariableExpansion(s):
 
   Returns:
       The escaped string.
-  """
+  """  # noqa: E731,E123,E501
     s = s.replace("%", "%%")
     return s
 
@@ -846,7 +835,7 @@ def _EscapeCppDefineForMSVS(s):
     s = _EscapeEnvironmentVariableExpansion(s)
     s = _EscapeCommandLineArgumentForMSVS(s)
     s = _EscapeVCProjCommandLineArgListItem(s)
-    # cl.exe replaces literal # characters with = in preprocesor definitions for
+    # cl.exe replaces literal # characters with = in preprocessor definitions for
     # some reason. Octal-encode to work around that.
     s = s.replace("#", "\\%03o" % ord("#"))
     return s
@@ -885,7 +874,7 @@ def _EscapeCppDefineForMSBuild(s):
     s = _EscapeEnvironmentVariableExpansion(s)
     s = _EscapeCommandLineArgumentForMSBuild(s)
     s = _EscapeMSBuildSpecialCharacters(s)
-    # cl.exe replaces literal # characters with = in preprocesor definitions for
+    # cl.exe replaces literal # characters with = in preprocessor definitions for
     # some reason. Octal-encode to work around that.
     s = s.replace("#", "\\%03o" % ord("#"))
     return s
@@ -1029,45 +1018,6 @@ def _GenerateProject(project, options, version, generator_flags, spec):
         return _GenerateMSVSProject(project, options, version, generator_flags)
 
 
-# TODO: Avoid code duplication with _ValidateSourcesForOSX in make.py.
-def _ValidateSourcesForMSVSProject(spec, version):
-    """Makes sure if duplicate basenames are not specified in the source list.
-
-  Arguments:
-    spec: The target dictionary containing the properties of the target.
-    version: The VisualStudioVersion object.
-  """
-    # This validation should not be applied to MSVC2010 and later.
-    assert not version.UsesVcxproj()
-
-    # TODO: Check if MSVC allows this for loadable_module targets.
-    if spec.get("type", None) not in ("static_library", "shared_library"):
-        return
-    sources = spec.get("sources", [])
-    basenames = {}
-    for source in sources:
-        name, ext = os.path.splitext(source)
-        is_compiled_file = ext in [".c", ".cc", ".cpp", ".cxx", ".m", ".mm", ".s", ".S"]
-        if not is_compiled_file:
-            continue
-        basename = os.path.basename(name)  # Don't include extension.
-        basenames.setdefault(basename, []).append(source)
-
-    error = ""
-    for basename, files in basenames.items():
-        if len(files) > 1:
-            error += "  %s: %s\n" % (basename, " ".join(files))
-
-    if error:
-        print(
-            "static library %s has several files with the same basename:\n"
-            % spec["target_name"]
-            + error
-            + "MSVC08 cannot handle that."
-        )
-        raise GypError("Duplicate basenames in sources section, see list above")
-
-
 def _GenerateMSVSProject(project, options, version, generator_flags):
     """Generates a .vcproj file.  It may create .rules and .user files too.
 
@@ -1093,11 +1043,6 @@ def _GenerateMSVSProject(project, options, version, generator_flags):
     config_type = _GetMSVSConfigurationType(spec, project.build_file)
     for config_name, config in spec["configurations"].items():
         _AddConfigurationToMSVSProject(p, spec, config_type, config_name, config)
-
-    # MSVC08 and prior version cannot handle duplicate basenames in the same
-    # target.
-    # TODO: Take excluded sources into consideration if possible.
-    _ValidateSourcesForMSVSProject(spec, version)
 
     # Prepare list of sources and excluded sources.
     gyp_file = os.path.split(project.build_file)[1]
@@ -1415,7 +1360,7 @@ def _GetOutputTargetExt(spec):
 
 
 def _GetDefines(config):
-    """Returns the list of preprocessor definitions for this configuation.
+    """Returns the list of preprocessor definitions for this configuration.
 
   Arguments:
     config: The dictionary that defines the special processing to be done
@@ -1620,7 +1565,7 @@ def _AdjustSourcesAndConvertToFilterHierarchy(
     if version.UsesVcxproj():
         while (
             all([isinstance(s, MSVSProject.Filter) for s in sources])
-            and len(set([s.name for s in sources])) == 1
+            and len({s.name for s in sources}) == 1
         ):
             assert all([len(s.contents) == 1 for s in sources])
             sources = [s.contents[0] for s in sources]
@@ -1826,8 +1771,8 @@ def _GetCopies(spec):
                 base_dir = posixpath.split(src_bare)[0]
                 outer_dir = posixpath.split(src_bare)[1]
                 fixed_dst = _FixPath(dst)
-                full_dst = '"%s\\%s\\"' % (fixed_dst, outer_dir)
-                cmd = 'mkdir %s 2>nul & cd "%s" && xcopy /e /f /y "%s" %s' % (
+                full_dst = f'"{fixed_dst}\\{outer_dir}\\"'
+                cmd = 'mkdir {} 2>nul & cd "{}" && xcopy /e /f /y "{}" {}'.format(
                     full_dst,
                     _FixPath(base_dir),
                     outer_dir,
@@ -1838,17 +1783,17 @@ def _GetCopies(spec):
                         [src],
                         ["dummy_copies", dst],
                         cmd,
-                        "Copying %s to %s" % (src, fixed_dst),
+                        f"Copying {src} to {fixed_dst}",
                     )
                 )
             else:
                 fix_dst = _FixPath(cpy["destination"])
-                cmd = 'mkdir "%s" 2>nul & set ERRORLEVEL=0 & copy /Y "%s" "%s"' % (
+                cmd = 'mkdir "{}" 2>nul & set ERRORLEVEL=0 & copy /Y "{}" "{}"'.format(
                     fix_dst,
                     _FixPath(src),
                     _FixPath(dst),
                 )
-                copies.append(([src], [dst], cmd, "Copying %s to %s" % (src, fix_dst)))
+                copies.append(([src], [dst], cmd, f"Copying {src} to {fix_dst}"))
     return copies
 
 
@@ -1948,12 +1893,12 @@ def _GetPlatformOverridesOfProject(spec):
     for config_name, c in spec["configurations"].items():
         config_fullname = _ConfigFullName(config_name, c)
         platform = c.get("msvs_target_platform", _ConfigPlatform(c))
-        fixed_config_fullname = "%s|%s" % (
+        fixed_config_fullname = "{}|{}".format(
             _ConfigBaseName(config_name, _ConfigPlatform(c)),
             platform,
         )
         if spec["toolset"] == "host" and generator_supports_multiple_toolsets:
-            fixed_config_fullname = "%s|x64" % (config_name,)
+            fixed_config_fullname = f"{config_name}|x64"
         config_platform_overrides[config_fullname] = fixed_config_fullname
     return config_platform_overrides
 
@@ -2106,7 +2051,7 @@ def PerformBuild(data, configurations, params):
 
     for config in configurations:
         arguments = [devenv, sln_path, "/Build", config]
-        print("Building [%s]: %s" % (config, arguments))
+        print(f"Building [{config}]: {arguments}")
         subprocess.check_call(arguments)
 
 
@@ -2221,7 +2166,12 @@ def GenerateOutput(target_list, target_dicts, data, params):
 
 
 def _GenerateMSBuildFiltersFile(
-    filters_path, source_files, rule_dependencies, extension_to_rule_name, platforms
+    filters_path,
+    source_files,
+    rule_dependencies,
+    extension_to_rule_name,
+    platforms,
+    toolset,
 ):
     """Generate the filters file.
 
@@ -2241,6 +2191,7 @@ def _GenerateMSBuildFiltersFile(
         rule_dependencies,
         extension_to_rule_name,
         platforms,
+        toolset,
         filter_group,
         source_group,
     )
@@ -2266,6 +2217,7 @@ def _AppendFiltersForMSBuild(
     rule_dependencies,
     extension_to_rule_name,
     platforms,
+    toolset,
     filter_group,
     source_group,
 ):
@@ -2277,7 +2229,7 @@ def _AppendFiltersForMSBuild(
       sources: The hierarchy of filters and sources to process.
       extension_to_rule_name: A dictionary mapping file extensions to rules.
       filter_group: The list to which filter entries will be appended.
-      source_group: The list to which source entries will be appeneded.
+      source_group: The list to which source entries will be appended.
   """
     for source in sources:
         if isinstance(source, MSVSProject.Filter):
@@ -2285,7 +2237,7 @@ def _AppendFiltersForMSBuild(
             if not parent_filter_name:
                 filter_name = source.name
             else:
-                filter_name = "%s\\%s" % (parent_filter_name, source.name)
+                filter_name = f"{parent_filter_name}\\{source.name}"
             # Add the filter to the group.
             filter_group.append(
                 [
@@ -2301,13 +2253,14 @@ def _AppendFiltersForMSBuild(
                 rule_dependencies,
                 extension_to_rule_name,
                 platforms,
+                toolset,
                 filter_group,
                 source_group,
             )
         else:
             # It's a source.  Create a source entry.
             _, element = _MapFileToMsBuildSourceType(
-                source, rule_dependencies, extension_to_rule_name, platforms
+                source, rule_dependencies, extension_to_rule_name, platforms, toolset
             )
             source_entry = [element, {"Include": source}]
             # Specify the filter it is part of, if any.
@@ -2317,7 +2270,7 @@ def _AppendFiltersForMSBuild(
 
 
 def _MapFileToMsBuildSourceType(
-    source, rule_dependencies, extension_to_rule_name, platforms
+    source, rule_dependencies, extension_to_rule_name, platforms, toolset
 ):
     """Returns the group and element type of the source file.
 
@@ -2345,9 +2298,8 @@ def _MapFileToMsBuildSourceType(
     elif ext in [".s", ".asm"]:
         group = "masm"
         element = "MASM"
-        for platform in platforms:
-            if platform.lower() in ["arm", "arm64"]:
-                element = "MARMASM"
+        if "arm64" in platforms and toolset == "target":
+            element = "MARMASM"
     elif ext == ".idl":
         group = "midl"
         element = "Midl"
@@ -2374,8 +2326,8 @@ def _GenerateRulesForMSBuild(
 ):
     # MSBuild rules are implemented using three files: an XML file, a .targets
     # file and a .props file.
-    # See http://blogs.msdn.com/b/vcblog/archive/2010/04/21/quick-help-on-vs2010-custom-build-rule.aspx
-    # for more details.
+    # For more details see:
+    # https://devblogs.microsoft.com/cppblog/quick-help-on-vs2010-custom-build-rule/
     rules = spec.get("rules", [])
     rules_native = [r for r in rules if not int(r.get("msvs_external_rule", 0))]
     rules_external = [r for r in rules if int(r.get("msvs_external_rule", 0))]
@@ -2413,7 +2365,7 @@ def _GenerateRulesForMSBuild(
     _AdjustSourcesForRules(rules, sources, excluded_sources, True)
 
 
-class MSBuildRule(object):
+class MSBuildRule:
     """Used to store information used to generate an MSBuild rule.
 
   Attributes:
@@ -2612,7 +2564,7 @@ def _GenerateMSBuildRuleTargetsFile(targets_path, msbuild_rules):
                 "Condition": "'@(%s)' != '' and '%%(%s.ExcludedFromBuild)' != "
                 "'true'" % (rule.tlog, rule.tlog),
                 "File": "$(IntDir)$(ProjectName).read.1.tlog",
-                "Lines": "^%%(%s.Source);%%(%s.Inputs)" % (rule.tlog, rule.tlog),
+                "Lines": f"^%({rule.tlog}.Source);%({rule.tlog}.Inputs)",
             },
         ]
         command_and_input_section = [
@@ -2958,7 +2910,7 @@ def _GetMSBuildProjectConfigurations(configurations, spec):
     group = ["ItemGroup", {"Label": "ProjectConfigurations"}]
     for (name, settings) in sorted(configurations.items()):
         configuration, platform = _GetConfigurationAndPlatform(name, settings, spec)
-        designation = "%s|%s" % (configuration, platform)
+        designation = f"{configuration}|{platform}"
         group.append(
             [
                 "ProjectConfiguration",
@@ -3323,13 +3275,11 @@ def _GetMSBuildPropertyGroup(spec, label, properties):
             # Self references are ignored. Self reference is used in a few places to
             # append to the default value. I.e. PATH=$(PATH);other_path
             edges.update(
-                set(
-                    [
-                        v
-                        for v in MSVS_VARIABLE_REFERENCE.findall(value)
-                        if v in properties and v != node
-                    ]
-                )
+                {
+                    v
+                    for v in MSVS_VARIABLE_REFERENCE.findall(value)
+                    if v in properties and v != node
+                }
             )
         return edges
 
@@ -3623,8 +3573,9 @@ def _AddSources2(
                     if precompiled_source != "":
                         precompiled_source = _FixPath(precompiled_source)
                         if not extensions_excluded_from_precompile:
-                            # If the precompiled header is generated by a C source, we must
-                            # not try to use it for C++ sources, and vice versa.
+                            # If the precompiled header is generated by a C source,
+                            # we must not try to use it for C++ sources,
+                            # and vice versa.
                             basename, extension = os.path.splitext(precompiled_source)
                             if extension == ".c":
                                 extensions_excluded_from_precompile = [
@@ -3656,7 +3607,18 @@ def _AddSources2(
                     rule_dependencies,
                     extension_to_rule_name,
                     _GetUniquePlatforms(spec),
+                    spec["toolset"],
                 )
+                if group == "compile" and not os.path.isabs(source):
+                    # Add an <ObjectFileName> value to support duplicate source
+                    # file basenames, except for absolute paths to avoid paths
+                    # with more than 260 characters.
+                    file_name = os.path.splitext(source)[0] + ".obj"
+                    if file_name.startswith("..\\"):
+                        file_name = re.sub(r"^(\.\.\\)+", "", file_name)
+                    elif file_name.startswith("$("):
+                        file_name = re.sub(r"^\$\([^)]+\)\\", "", file_name)
+                    detail.append(["ObjectFileName", "$(IntDir)\\" + file_name])
                 grouped_sources[group].append([element, {"Include": source}] + detail)
 
 
@@ -3704,6 +3666,7 @@ def _GetMSBuildProjectReferences(project):
 def _GenerateMSBuildProject(project, options, version, generator_flags, spec):
     spec = project.spec
     configurations = spec["configurations"]
+    toolset = spec["toolset"]
     project_dir, project_file_name = os.path.split(project.path)
     gyp.common.EnsureDirExists(project.path)
     # Prepare list of sources and excluded sources.
@@ -3717,6 +3680,7 @@ def _GenerateMSBuildProject(project, options, version, generator_flags, spec):
     rule_dependencies = set()
     extension_to_rule_name = {}
     list_excluded = generator_flags.get("msvs_list_excluded_files", True)
+    platforms = _GetUniquePlatforms(spec)
 
     # Don't generate rules if we are using an external builder like ninja.
     if not spec.get("msvs_external_builder"):
@@ -3759,7 +3723,8 @@ def _GenerateMSBuildProject(project, options, version, generator_flags, spec):
         sources,
         rule_dependencies,
         extension_to_rule_name,
-        _GetUniquePlatforms(spec),
+        platforms,
+        toolset,
     )
     missing_sources = _VerifySourcesExist(sources, project_dir)
 
@@ -3812,7 +3777,7 @@ def _GenerateMSBuildProject(project, options, version, generator_flags, spec):
         content += _GetMSBuildLocalProperties(project.msbuild_toolset)
     content += import_cpp_props_section
     content += import_masm_props_section
-    if spec.get("msvs_enable_marmasm"):
+    if "arm64" in platforms and toolset == "target":
         content += import_marmasm_props_section
     content += _GetMSBuildExtensions(props_files_of_rules)
     content += _GetMSBuildPropertySheets(configurations, spec)
@@ -3834,7 +3799,7 @@ def _GenerateMSBuildProject(project, options, version, generator_flags, spec):
     content += _GetMSBuildProjectReferences(project)
     content += import_cpp_targets_section
     content += import_masm_targets_section
-    if spec.get("msvs_enable_marmasm"):
+    if "arm64" in platforms and toolset == "target":
         content += import_marmasm_targets_section
     content += _GetMSBuildExtensionTargets(targets_files_of_rules)
 
