@@ -2,7 +2,6 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-from __future__ import print_function
 
 import filecmp
 import gyp.common
@@ -110,7 +109,7 @@ def CreateXCConfigurationList(configuration_names):
     return xccl
 
 
-class XcodeProject(object):
+class XcodeProject:
     def __init__(self, gyp_path, path, build_file_dict):
         self.gyp_path = gyp_path
         self.path = path
@@ -594,7 +593,7 @@ _xcode_define_re = re.compile(r"([\\\"\' ])")
 def EscapeXcodeDefine(s):
     """We must escape the defines that we give to XCode so that it knows not to
      split on spaces and to respect backslash and quote literals. However, we
-     must not quote the define, or Xcode will incorrectly intepret variables
+     must not quote the define, or Xcode will incorrectly interpret variables
      especially $(inherited)."""
     return re.sub(_xcode_define_re, r"\\\1", s)
 
@@ -613,7 +612,7 @@ def PerformBuild(data, configurations, params):
     for config in configurations:
         arguments = ["xcodebuild", "-project", xcodeproj_path]
         arguments += ["-configuration", config]
-        print("Building [%s]: %s" % (config, arguments))
+        print(f"Building [{config}]: {arguments}")
         subprocess.check_call(arguments)
 
 
@@ -735,7 +734,8 @@ def GenerateOutput(target_list, target_dicts, data, params):
             "loadable_module+xcuitest": "com.apple.product-type.bundle.ui-testing",
             "shared_library+bundle": "com.apple.product-type.framework",
             "executable+extension+bundle": "com.apple.product-type.app-extension",
-            "executable+watch+extension+bundle": "com.apple.product-type.watchkit-extension",
+            "executable+watch+extension+bundle":
+                "com.apple.product-type.watchkit-extension",
             "executable+watch+bundle": "com.apple.product-type.application.watchapp",
             "mac_kernel_extension+bundle": "com.apple.product-type.kernel-extension",
         }
@@ -1026,7 +1026,7 @@ def GenerateOutput(target_list, target_dicts, data, params):
                 for output in rule.get("outputs", []):
                     # Fortunately, Xcode and make both use $(VAR) format for their
                     # variables, so the expansion is the only transformation necessary.
-                    # Any remaning $(VAR)-type variables in the string can be given
+                    # Any remaining $(VAR)-type variables in the string can be given
                     # directly to make, which will pick up the correct settings from
                     # what Xcode puts into the environment.
                     concrete_output = ExpandXcodeVariables(output, rule_input_dict)
@@ -1071,7 +1071,7 @@ def GenerateOutput(target_list, target_dicts, data, params):
                 # TODO(mark): There's a possibility for collision here.  Consider
                 # target "t" rule "A_r" and target "t_A" rule "r".
                 makefile_name = "%s.make" % re.sub(
-                    "[^a-zA-Z0-9_]", "_", "%s_%s" % (target_name, rule["rule_name"])
+                    "[^a-zA-Z0-9_]", "_", "{}_{}".format(target_name, rule["rule_name"])
                 )
                 makefile_path = os.path.join(
                     xcode_projects[build_file].path, makefile_name
@@ -1101,7 +1101,7 @@ def GenerateOutput(target_list, target_dicts, data, params):
                         eol = ""
                     else:
                         eol = " \\"
-                    makefile.write("    %s%s\n" % (concrete_output, eol))
+                    makefile.write(f"    {concrete_output}{eol}\n")
 
                 for (rule_source, concrete_outputs, message, action) in zip(
                     rule["rule_sources"],
@@ -1122,7 +1122,7 @@ def GenerateOutput(target_list, target_dicts, data, params):
                             bol = ""
                         else:
                             bol = "    "
-                        makefile.write("%s%s \\\n" % (bol, concrete_output))
+                        makefile.write(f"{bol}{concrete_output} \\\n")
 
                         concrete_output_dir = posixpath.dirname(concrete_output)
                         if (
@@ -1142,7 +1142,7 @@ def GenerateOutput(target_list, target_dicts, data, params):
                             eol = ""
                         else:
                             eol = " \\"
-                        makefile.write("    %s%s\n" % (prerequisite, eol))
+                        makefile.write(f"    {prerequisite}{eol}\n")
 
                     # Make sure that output directories exist before executing the rule
                     # action.
@@ -1151,8 +1151,8 @@ def GenerateOutput(target_list, target_dicts, data, params):
                             '\t@mkdir -p "%s"\n' % '" "'.join(concrete_output_dirs)
                         )
 
-                    # The rule message and action have already had the necessary variable
-                    # substitutions performed.
+                    # The rule message and action have already had
+                    # the necessary variable substitutions performed.
                     if message:
                         # Mark it with note: so Xcode picks it up in build output.
                         makefile.write("\t@echo note: %s\n" % message)
@@ -1204,8 +1204,8 @@ exit 1
                     support_xct.AppendProperty("buildPhases", ssbp)
                 else:
                     # TODO(mark): this assumes too much knowledge of the internals of
-                    # xcodeproj_file; some of these smarts should move into xcodeproj_file
-                    # itself.
+                    # xcodeproj_file; some of these smarts should move
+                    # into xcodeproj_file itself.
                     xct._properties["buildPhases"].insert(prebuild_index, ssbp)
                     prebuild_index = prebuild_index + 1
 
