@@ -2,7 +2,6 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-from __future__ import print_function
 
 import ast
 
@@ -20,8 +19,6 @@ import traceback
 from distutils.version import StrictVersion
 from gyp.common import GypError
 from gyp.common import OrderedSet
-
-PY3 = bytes != str
 
 # A list of types that are treated as linkable.
 linkable_types = [
@@ -66,7 +63,7 @@ def IsPathSection(section):
     if section in path_sections:
         return True
 
-    # Sections mathing the regexp '_(dir|file|path)s?$' are also
+    # Sections matching the regexp '_(dir|file|path)s?$' are also
     # considered PathSections. Using manual string matching since that
     # is much faster than the regexp and this can be called hundreds of
     # thousands of times so micro performance matters.
@@ -228,17 +225,9 @@ def LoadOneBuildFile(build_file_path, data, aux_data, includes, is_target, check
         return data[build_file_path]
 
     if os.path.exists(build_file_path):
-        # Open the build file for read ('r') with universal-newlines mode ('U')
-        # to make sure platform specific newlines ('\r\n' or '\r') are converted to '\n'
-        # which otherwise will fail eval()
-        if sys.platform == "zos":
-            # On z/OS, universal-newlines mode treats the file as an ascii file. But since
-            # node-gyp produces ebcdic files, do not use that mode.
-            build_file_contents = open(build_file_path, "r").read()
-        else:
-            build_file_contents = open(build_file_path, "rU").read()
+        build_file_contents = open(build_file_path).read()
     else:
-        raise GypError("%s not found (cwd: %s)" % (build_file_path, os.getcwd()))
+        raise GypError(f"{build_file_path} not found (cwd: {os.getcwd()})")
 
     build_file_data = None
     try:
@@ -567,7 +556,7 @@ class ParallelProcessingError(Exception):
     pass
 
 
-class ParallelState(object):
+class ParallelState:
     """Class to keep track of state when processing input files in parallel.
 
   If build files are loaded in parallel, use this to keep track of
@@ -987,9 +976,8 @@ def ExpandVariables(input, phase, variables, build_file):
                         )
 
                     p_stdout, p_stderr = p.communicate("")
-                    if PY3:
-                        p_stdout = p_stdout.decode("utf-8")
-                        p_stderr = p_stderr.decode("utf-8")
+                    p_stdout = p_stdout.decode("utf-8")
+                    p_stderr = p_stderr.decode("utf-8")
 
                     if p.wait() != 0 or p_stderr:
                         sys.stderr.write(p_stderr)
@@ -1219,7 +1207,7 @@ def EvalSingleCondition(cond_expr, true_dict, false_dict, phase, variables, buil
     except NameError as e:
         gyp.common.ExceptionAppend(
             e,
-            "while evaluating condition '%s' in %s" % (cond_expr_expanded, build_file),
+            f"while evaluating condition '{cond_expr_expanded}' in {build_file}",
         )
         raise GypError(e)
 
@@ -1675,7 +1663,7 @@ def RemoveLinkDependenciesFromNoneTargets(targets):
                             )
 
 
-class DependencyGraphNode(object):
+class DependencyGraphNode:
     """
 
   Attributes:
@@ -1731,10 +1719,10 @@ class DependencyGraphNode(object):
                     node_dependent.dependencies, key=ExtractNodeRef
                 ):
                     if node_dependent_dependency.ref not in flat_list:
-                        # The dependent one or more dependencies not in flat_list.  There
-                        # will be more chances to add it to flat_list when examining
-                        # it again as a dependent of those other dependencies, provided
-                        # that there are no cycles.
+                        # The dependent one or more dependencies not in flat_list.
+                        # There will be more chances to add it to flat_list
+                        # when examining it again as a dependent of those other
+                        # dependencies, provided that there are no cycles.
                         is_in_degree_zero = False
                         break
 
@@ -2252,7 +2240,7 @@ def MergeLists(to, fro, to_file, fro_file, is_paths=False, append=True):
 
     # Make membership testing of hashables in |to| (in particular, strings)
     # faster.
-    hashable_to_set = set(x for x in to if is_hashable(x))
+    hashable_to_set = {x for x in to if is_hashable(x)}
     for item in fro:
         singleton = False
         if type(item) in (str, int):
@@ -2427,7 +2415,7 @@ def MergeDicts(to, fro, to_file, fro_file):
 def MergeConfigWithInheritance(
     new_configuration_dict, build_file, target_dict, configuration, visited
 ):
-    # Skip if previously visted.
+    # Skip if previously visited.
     if configuration in visited:
         return
 
@@ -2641,10 +2629,10 @@ def ProcessListFiltersInDict(name, the_dict):
                 pattern_re = re.compile(pattern)
 
                 if action == "exclude":
-                    # This item matches an exclude regex, so set its value to 0 (exclude).
+                    # This item matches an exclude regex, set its value to 0 (exclude).
                     action_value = 0
                 elif action == "include":
-                    # This item matches an include regex, so set its value to 1 (include).
+                    # This item matches an include regex, set its value to 1 (include).
                     action_value = 1
                 else:
                     # This is an action that doesn't make any sense.
@@ -2659,8 +2647,8 @@ def ProcessListFiltersInDict(name, the_dict):
 
                 for index, list_item in enumerate(the_list):
                     if list_actions[index] == action_value:
-                        # Even if the regex matches, nothing will change so continue (regex
-                        # searches are expensive).
+                        # Even if the regex matches, nothing will change so continue
+                        # (regex searches are expensive).
                         continue
                     if pattern_re.search(list_item):
                         # Regular expression match.
@@ -2750,36 +2738,6 @@ def ValidateTargetType(target, target_dict):
         )
 
 
-def ValidateSourcesInTarget(target, target_dict, build_file, duplicate_basename_check):
-    if not duplicate_basename_check:
-        return
-    if target_dict.get("type", None) != "static_library":
-        return
-    sources = target_dict.get("sources", [])
-    basenames = {}
-    for source in sources:
-        name, ext = os.path.splitext(source)
-        is_compiled_file = ext in [".c", ".cc", ".cpp", ".cxx", ".m", ".mm", ".s", ".S"]
-        if not is_compiled_file:
-            continue
-        basename = os.path.basename(name)  # Don't include extension.
-        basenames.setdefault(basename, []).append(source)
-
-    error = ""
-    for basename, files in basenames.items():
-        if len(files) > 1:
-            error += "  %s: %s\n" % (basename, " ".join(files))
-
-    if error:
-        print(
-            "static library %s has several files with the same basename:\n" % target
-            + error
-            + "libtool on Mac cannot handle that. Use "
-            "--no-duplicate-basename-check to disable this validation."
-        )
-        raise GypError("Duplicate basenames in sources section, see list above")
-
-
 def ValidateRulesInTarget(target, target_dict, extra_sources_for_rules):
     """Ensures that the rules sections in target_dict are valid and consistent,
   and determines which sources they apply to.
@@ -2802,7 +2760,7 @@ def ValidateRulesInTarget(target, target_dict, extra_sources_for_rules):
         rule_name = rule["rule_name"]
         if rule_name in rule_names:
             raise GypError(
-                "rule %s exists in duplicate, target %s" % (rule_name, target)
+                f"rule {rule_name} exists in duplicate, target {target}"
             )
         rule_names[rule_name] = rule
 
@@ -3021,7 +2979,6 @@ def Load(
     generator_input_info,
     check,
     circular_check,
-    duplicate_basename_check,
     parallel,
     root_targets,
 ):
@@ -3167,9 +3124,6 @@ def Load(
         target_dict = targets[target]
         build_file = gyp.common.BuildFile(target)
         ValidateTargetType(target, target_dict)
-        ValidateSourcesInTarget(
-            target, target_dict, build_file, duplicate_basename_check
-        )
         ValidateRulesInTarget(target, target_dict, extra_sources_for_rules)
         ValidateRunAsInTarget(target, target_dict, build_file)
         ValidateActionsInTarget(target, target_dict, build_file)
