@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 # Copyright (c) 2012 Google Inc. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be
@@ -9,7 +9,6 @@
 These functions are executed via gyp-win-tool when using the ninja generator.
 """
 
-from __future__ import print_function
 
 import os
 import re
@@ -20,7 +19,6 @@ import string
 import sys
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-PY3 = bytes != str
 
 # A regex matching an argument corresponding to the output filename passed to
 # link.exe.
@@ -34,7 +32,7 @@ def main(args):
         sys.exit(exit_code)
 
 
-class WinTool(object):
+class WinTool:
     """This class performs all the Windows tooling steps. The methods can either
   be executed directly, or dispatched from an argument list."""
 
@@ -141,9 +139,7 @@ class WinTool(object):
             stdout=subprocess.PIPE,
             stderr=subprocess.STDOUT,
         )
-        out, _ = link.communicate()
-        if PY3:
-            out = out.decode("utf-8")
+        out = link.communicate()[0].decode("utf-8")
         for line in out.splitlines():
             if (
                 not line.startswith("   Creating library ")
@@ -223,8 +219,8 @@ class WinTool(object):
             our_manifest = "%(out)s.manifest" % variables
             # Load and normalize the manifests. mt.exe sometimes removes whitespace,
             # and sometimes doesn't unfortunately.
-            with open(our_manifest, "r") as our_f:
-                with open(assert_manifest, "r") as assert_f:
+            with open(our_manifest) as our_f:
+                with open(assert_manifest) as assert_f:
                     our_data = our_f.read().translate(None, string.whitespace)
                     assert_data = assert_f.read().translate(None, string.whitespace)
             if our_data != assert_data:
@@ -233,7 +229,7 @@ class WinTool(object):
                 def dump(filename):
                     print(filename, file=sys.stderr)
                     print("-----", file=sys.stderr)
-                    with open(filename, "r") as f:
+                    with open(filename) as f:
                         print(f.read(), file=sys.stderr)
                         print("-----", file=sys.stderr)
 
@@ -256,9 +252,7 @@ class WinTool(object):
         popen = subprocess.Popen(
             args, shell=True, env=env, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
         )
-        out, _ = popen.communicate()
-        if PY3:
-            out = out.decode("utf-8")
+        out = popen.communicate()[0].decode("utf-8")
         for line in out.splitlines():
             if line and "manifest authoring warning 81010002" not in line:
                 print(line)
@@ -302,16 +296,14 @@ class WinTool(object):
         popen = subprocess.Popen(
             args, shell=True, env=env, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
         )
-        out, _ = popen.communicate()
-        if PY3:
-            out = out.decode("utf-8")
+        out = popen.communicate()[0].decode("utf-8")
         # Filter junk out of stdout, and write filtered versions. Output we want
         # to filter is pairs of lines that look like this:
         # Processing C:\Program Files (x86)\Microsoft SDKs\...\include\objidl.idl
         # objidl.idl
         lines = out.splitlines()
         prefixes = ("Processing ", "64 bit Processing ")
-        processing = set(os.path.basename(x) for x in lines if x.startswith(prefixes))
+        processing = {os.path.basename(x) for x in lines if x.startswith(prefixes)}
         for line in lines:
             if not line.startswith(prefixes) and line not in processing:
                 print(line)
@@ -323,9 +315,7 @@ class WinTool(object):
         popen = subprocess.Popen(
             args, shell=True, env=env, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
         )
-        out, _ = popen.communicate()
-        if PY3:
-            out = out.decode("utf-8")
+        out = popen.communicate()[0].decode("utf-8")
         for line in out.splitlines():
             if (
                 not line.startswith("Copyright (C) Microsoft Corporation")
@@ -343,9 +333,7 @@ class WinTool(object):
         popen = subprocess.Popen(
             args, shell=True, env=env, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
         )
-        out, _ = popen.communicate()
-        if PY3:
-            out = out.decode("utf-8")
+        out = popen.communicate()[0].decode("utf-8")
         for line in out.splitlines():
             if (
                 not line.startswith("Microsoft (R) Windows (R) Resource Compiler")
