@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 # Copyright (c) 2012 Google Inc. All rights reserved.
 # Use of this source code is governed by a BSD-style license that can be
@@ -9,10 +9,7 @@
 import unittest
 import gyp.MSVSSettings as MSVSSettings
 
-try:
-    from StringIO import StringIO  # Python 2
-except ImportError:
-    from io import StringIO  # Python 3
+from io import StringIO
 
 
 class TestSequenceFunctions(unittest.TestCase):
@@ -678,7 +675,8 @@ class TestSequenceFunctions(unittest.TestCase):
                 "MSBuild, index value (21) not in expected range [0, 3)",
                 "Warning: while converting VCCLCompilerTool/UsePrecompiledHeader to "
                 "MSBuild, index value (13) not in expected range [0, 3)",
-                "Warning: while converting VCCLCompilerTool/GeneratePreprocessedFile to "
+                "Warning: while converting "
+                "VCCLCompilerTool/GeneratePreprocessedFile to "
                 "MSBuild, value must be one of [0, 1, 2]; got 14",
                 "Warning: while converting VCLinkerTool/Driver to "
                 "MSBuild, index value (10) not in expected range [0, 4)",
@@ -1348,7 +1346,8 @@ class TestSequenceFunctions(unittest.TestCase):
                 "EmbedManifest": "false",
                 "GenerateCatalogFiles": "true",
                 "InputResourceManifests": "asfsfdafs",
-                "ManifestResourceFile": "$(IntDir)\\$(TargetFileName).embed.manifest.resfdsf",
+                "ManifestResourceFile":
+                    "$(IntDir)\\$(TargetFileName).embed.manifest.resfdsf",
                 "OutputManifestFile": "$(TargetPath).manifestdfs",
                 "RegistrarScriptFile": "sdfsfd",
                 "ReplacementsFile": "sdffsd",
@@ -1532,7 +1531,8 @@ class TestSequenceFunctions(unittest.TestCase):
                 "LinkIncremental": "",
             },
             "ManifestResourceCompile": {
-                "ResourceOutputFileName": "$(IntDir)$(TargetFileName).embed.manifest.resfdsf"
+                "ResourceOutputFileName":
+                    "$(IntDir)$(TargetFileName).embed.manifest.resfdsf"
             },
         }
         self.maxDiff = 9999  # on failure display a long diff
